dialog-changes

+++ src/app/products/products.component.ts
@@ -6,6 +6,11 @@ import { MatTableDataSource } from '@angular/material/table';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSort, Sort } from '@angular/material/sort';
 import { LiveAnnouncer } from '@angular/cdk/a11y';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
+import { analyzeAndValidateNgModules } from '@angular/compiler';
+
+
 
 export interface Product {
   id: number;
@@ -38,7 +43,8 @@ export class ProductsComponent implements OnInit, AfterViewInit {
     private dataService: DataService,
     private route: ActivatedRoute,
     private router: Router,
-    private _liveAnnouncer: LiveAnnouncer
+    private _liveAnnouncer: LiveAnnouncer,
+    private dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -58,23 +64,34 @@ export class ProductsComponent implements OnInit, AfterViewInit {
   }
 
   deleteProduct(id) {
-    this.dataService.deleteProduct(id).subscribe((response) => {
-      console.log(response);
-      this.loadData();
+   
+      const confirmDialog = this.dialog.open(ConfirmDialogComponent, {
+        data: {
+          title: 'Confirm Remove Employee',
+          message: 'Are you sure, you want to remove an employee: '+ id        
+        }
+      
       // this.router.navigate(['/products']);
     });
-  }
-
+    confirmDialog.afterClosed().subscribe(result => {
+      if (result === true) {
+      this.dataService.deleteProduct(id).subscribe((response) => {
+        console.log(response);
+        this.loadData();
+      
+    });
+  
+}}
   /** Announce the change in sort state for assistive technology. */
-  announceSortChange(sortState: any) {
-    // This example uses English messages. If your application supports
-    // multiple language, you would internationalize these strings.
-    // Furthermore, you can customize the message to add additional
-    // details about the values being sorted.
-    if (sortState.direction) {
-      this._liveAnnouncer.announce(`Sorted ${sortState.direction}ending`);
-    } else {
-      this._liveAnnouncer.announce('Sorting cleared');
-    }
-  }
-}
+  // announceSortChange(sortState: any) {
+//     // This example uses English messages. If your application supports
+//     // multiple language, you would internationalize these strings.
+//     // Furthermore, you can customize the message to add additional
+//     // details about the values being sorted.
+//     if (sortState.direction) {
+//       this._liveAnnouncer.announce(`Sorted ${sortState.direction}ending`);
+//     } else {
+//       this._liveAnnouncer.announce('Sorting cleared');
+//     }
+//   }
+// }
